requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.5
asttokens==2.1.0
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.4.0
backcall==0.2.0
chardet==4.0.0
decorator==5.1.1
discord.py==1.7.3
executing==1.2.0
greenlet==1.1.2
idna==3.3
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
multidict==5.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
stack-data==0.6.0
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.0.1
wcwidth==0.2.5
yarl==1.7.2
